requirements.txt
sentence-transformers>=2.2.0
numpy>=1.21.0
tiktoken>=0.5.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0
PyYAML>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
